✨ feat(eval): let a dataset exist without a benchmark
A benchmark models a *published* suite: it carries `rubrics`, a `referenceUrl`
and `isSystem` defaulting to true. A dataset accumulated from cases captured out
of real conversations has none of that — each of its cases carries its own
criteria — yet `agent_eval_datasets.benchmark_id` was `NOT NULL`, so such a
dataset had to invent a fake benchmark row to exist. That fiction also put the
captured cases under the benchmark's delete cascade, where deleting the
placeholder would take them with it.

Makes the column nullable, and follows the consequence through:

- Scoring no longer bails when a dataset has no benchmark. Both evaluation paths
  did `findById(dataset.benchmarkId)` and returned early if absent; left as-is,
  every case in a benchmark-less dataset would have been silently unscored.
  A missing benchmark now simply contributes no benchmark-level rubrics, and
  scoring falls back to the per-case / per-dataset `evalMode` — which is exactly
  what a captured case uses.
- A dataset and a case get canonical addresses that do not carry a benchmark id:
  `/eval/datasets/:datasetId` and `/eval/cases/:caseId`. The previous nested
  path could not even be constructed for a benchmark-less dataset. The old
  `/eval/bench/:benchmarkId/datasets/:datasetId` redirects, so existing links
  keep resolving.
- `createDataset` accepts no benchmark; `listDatasets({ unlinked: true })`
  returns the ones that have none.

Runs are deliberately not migrated: every run URL is still
`/eval/bench/:benchmarkId/runs/:runId`, and moving them is an IA change (the
benchmark sidebar is currently the only eval chrome), not a rename. The dataset
page therefore gates its runs section on having a benchmark and says so, rather
than rendering links with a null benchmark in the path.
