+++ locales/en-US/eval.json
@@ -131,8 +131,10 @@
   "dataset.delete.success": "Dataset deleted successfully",
   "dataset.detail.addRun": "New Run",
   "dataset.detail.backToBenchmark": "Back to Benchmark",
+  "dataset.detail.backToEval": "Back to Eval",
   "dataset.detail.caseCount": "{{count}} test case{{count, plural, one {} other {s}}}",
   "dataset.detail.relatedRuns": "Related Runs ({{count}})",
+  "dataset.detail.runsNeedBenchmark": "Evaluation runs are still scoped to a benchmark, so this dataset cannot be run yet.",
   "dataset.detail.testCases": "Test Cases",
   "dataset.detail.viewDetail": "View Details",
   "dataset.edit.error": "Failed to update dataset",
