+++ locales/zh-CN/eval.json
@@ -131,8 +131,10 @@
   "dataset.delete.success": "数据集删除成功",
   "dataset.detail.addRun": "新增运行",
   "dataset.detail.backToBenchmark": "返回基准测试",
+  "dataset.detail.backToEval": "返回评测",
   "dataset.detail.caseCount": "{{count}} 个测试用例",
   "dataset.detail.relatedRuns": "相关运行（{{count}}）",
+  "dataset.detail.runsNeedBenchmark": "评测运行目前仍归属于某个基准测试，因此该数据集暂时无法发起运行。",
   "dataset.detail.testCases": "测试用例",
   "dataset.detail.viewDetail": "查看详情",
   "dataset.edit.error": "更新数据集失败",
